requirements.txt
aiohttp>=3.7.4
asyncio==3.4.3
gitpython==3.1.36
heroku3==5.2.1
pyrogram==2.0.106
python-dotenv==1.0.0
requests==2.31.0
TgCrypto==1.2.5
urllib3==2.0.4
